Move `vertex` into its own module
